Update: 描画ロジックを分離
